refactor: Standardize votingEndTimestamp to uint48

+++ contracts/deployables/strategies/LinearERC721VotingV1.sol
@@ -509,7 +509,7 @@ contract LinearERC721VotingV1 is
                 emit VotingPeriodEnded(
                     _proposalId,
                     proposal.votingEndTimestamp,
-                    block.timestamp
+                    uint48(block.timestamp)
                 );
                 return;
             }
